+++ src/vector2f.cpp
@@ -1,7 +1,7 @@
 #include "../include/vector2f.hpp"
 #include <iostream>
 
-Vector2f::Vector2f(float p_x, float p_y)
+Vector2f::Vector2f(int p_x, int p_y)
 :x(p_x), y(p_y)
 {}
 
@@ -10,12 +10,12 @@ void Vector2f::print()
 	std::cout << x << ", " << y << std::endl;
 }
 
-float Vector2f::getX()
+int Vector2f::getX()
 {
 	return x;
 }
 
-float Vector2f::getY()
+int Vector2f::getY()
 {
 	return y;
 }
